Auto-provision new Users based on data in the OpenToken.
If configured to do so, new Users are *automatically* provisioned based on the
data that has been provided to us in the OpenToken.

This allows for an organization to use SAML requests to not only authenticate
their Users for their Socialtext Appliance, but also allows for Users to be
provisioned "on the fly" as they access the system.