Specify unit test fixtures, for test run optimization.
Our test infrastructure doesn't pick up on "fixtures generated out of process"
when it tries to optimize the order of the unit tests in a test run.  Help it
out by explicitly specifying the fixtures (being that they've just been
generated, having the line here will _not_ cause it to get generated a second
time; it'll be re-used instead).