Add a new "/nlw/logout.html" template.
This template can be used to indicate to Users that they have been logged out
of the system.

This template does *NOT* contain a "login form" on it on purpose; not all of
our customers use our standard login form and don't want to imply to Users
that they can enter their user/pass here to login.

We are *not* changing the default redirect of where you get sent to on logout;
that still remains "/nlw/login.html".  This template merely provides another
_option_ for a page to show to the User.