Fix {bz: 3586}; *resolve* Users, don't just look them up.
After much going back and forth with the customer, e3vil managed to figure out
what the chain of events was that caused the error to be displayed on page views.

Boiled down to ... if the "username" and "email address" for a Default User
got out of sync, any pages that were created or last edited by that User
*before* things got out of sync were now completely inaccessible.  Attempting
to access those Pages (or view Page history on a page that they've edited in
the past) would cause an error to be displayed in the browser.

If, however, we *resolve* the User records (versus just looking them up by
e-mail address directly), we find the User record that exists in the DB and
don't trigger the "this User is gone, go create a placeholder User record for
them" code.

Yes, I know that there's a whole other question around "how did these Users
end up getting their e-mail/username out of sync in the first place?", but
when this happens we should provide better behaviour than blowing up.