Stop/start console supervisor during fdefs.
If we don't explicitly stop/start the console supervisor during fdefs, its
easy to get multiple copies of it running (just nuke ~/.nlw/ between fdefs and
you've got multiple copies running).

This is totally reminiscent of problems we had in the past with
st-bootstrap-openldap in that once ~/.nlw/ is nuked we lose access to any PID
files that might have been in there and as a result we aren't able to detect
that we have running processes that need to be killed off.

With the console-supervisor its more fun, though, as the PID file we have
*isn't* for the actual "supervise" process itself but is for the wrapper.
Further, App::Daemon isn't able to kill off the running supervisor on its own
and so if we need to stop it then we have to use "svc" to do that separately.