Display "comments" in e-mail notifications.
Yes, I am aware that this change *appears* to undo the "| html" escaping on
the edit summary.  "Appears to" is the right phrase, though... the syntax that
was being used actually *never* escaped things in the first place; the summary
was being escaped when it was being placed in the DB, not here.