Show dms when both users are a member of the filtered network
When you're looking at a particular group or account's events, you
don't want to see all direct messages between you and everyone. Instead,
you wnat to filter that list to users in the filtered network. In order
to do this, we need to ignore the signal_user_set table for direct
messages, and instead just check that both the actor_id and the
person_id columns refer to a user that is a user in the current network.