Start of a unit-test to check for bad deeplinks to "/nlw/login.html"
We really shouldn't be deeplinking directly to the "login template" when we
want to tell the User that they need to login.  Instead, should actually be
triggering the "Challenger" sub-system (which then DTRT to log the person in)
(and which by default shows the login template).

So... keep an eye out for places that we use the login template directly, and
complain about them; we really should be doing something else there.