Change "default_logout_uri" to be "the Challenger" and not "the login form".
In a standard ST configuration, this'll behave the same as it did before;
Users will click "logout" and will be taken back to a login form.

In a SSO configuration, this will re-trigger the Challenger, which will then
log the User back into Socialtext again.  Yes, seems odd... *BUT*... in an SSO
configuration even if the User logs out of ST they're still logged in to their
SSO system (and so the very next ST URI they hit would automatically log them
back in _anyways_).