/nlw/login.html exception... script contains test data.
This script contains *raw* Apache log data which it then feeds in to create
some basic reporting data.  As such, its going to have links in it to the
login template, so make sure that we exclude this from our test of "nothing
should refer to the login template directly".