Two things to fix this test...
1) We need to make certain that we're running against a clean db/fs

2) When we pass in a 'date' param to ST::Page->create(), we assume that the
date is in UTC, which then gets re-written to localtime. For purposes of this
test, midnight, January 1, 2003 gets translated to 4pm, December 31, 2002 (for
example), which causes breakage. Since the only time we appear to be passing
an explicit date param in to this sub is in the test, we're safe just changing
the datetime to midnight, January 2, 2003.