Fix {bz: 3717}; Socialtext.User needs a "user_id", not a "username".
Bug is that when we have an all-numeric username, URLs in the form of
"/data/users/00012" get treated by the back-end as "user_id 12" and not
"username 00012".

Instead of leaving ourselves open to this misdirection, create the URL in the
form of "/data/users/<user_id>" instead of ".../<username>".  We *know* that
the user_id is going to be the first case looked up by ST::User->Resolve()
when handed something that looks numeric.

wikiQ test tacked onto the test for {bz: 3708} as its got immediate locality;
both issues are on the same page in the CP and are stemmed from the same cause.