Update test for new default pages.
The pages included in a newly created workspace changed in the BK branch.
Update these tests to reflect those changes.

We should probably rewrite this test at some point to make fewer assumptions
about what data is and is not present by default, however, this late in a
release cycle is _not_ that time.