Addresses {bz: 3717}; Workspace trash should get Users to delete by "user_id" first, then "username".
If we have an all-numeric username (e.g. "username: 00012") that causes
potential problems when their username conflicts with the user_id of another
User (e.g. "user_id: 12").  When we're gathering up the list of Users to
remove, look for a "user_id" first as we know that one is guaranteed to be
unique and not create conflict.