Add template and helper method to show error messages.
Template was based on "authen/login.html", but edited so that we *don't* show
any login entry fields on the page; idea here is that we're showing an error
message (but that its not one that you can necessarily fix just by logging
in) (e.g. its not "You must login to view this page").