Display error page (not login page) if we can't find the User mentioned in the confirmation hash.
If we can't find the User, the error page displayed _shouldn't_ show the login
fields on it but intsead should *just* be an error page.

Note that "_find_user_for_email_confirmation_hash()" sets an appropriate error
message into the session, so we don't have to do it ourselves or pass it
through to "_show_error()"; it'll just get picked up from the session.