If we can't resend the confirmation e-mail, show the User _something_.
Otherwise the User has just clicked on the "Resend Confirmation" button and
isn't getting any indication of what's gone on or why he/she isn't going to
get something in their e-mail.

Admittedly we _shouldn't_ ever hit this, but if we do it'd be nice to at least
tell the User why they're not going to get anything in their Inbox.