Display an error to the User if we're missing the confirmation hash.
Rather than throwing the User back to the login page with *no* indication of
what has gone wrong, show them an error page indicating that the URL they've
clicked on to get here is invalid.  Not its not _that_ much better, but it at
least gives Users some indication of what's gone wrong.