Ignore the "t/" directory entirely; tests _are_ likely to contain references to "/nlw/login.html".
We have *several* places internally where we trigger the challenger *inline*
as opposed to redirecting to "/challenge".  When this happens, the HTTP
response will redirect _directly_ to "/nlw/login.html".  As such, its likely
that this string will appear within unit/wikiD tests.  So, ignore the test
directory when searching for potential mis-uses of it as a string.