Fix {bz: 3757}; Don't show the "you need to login" message on the login page; its redundant.
This (unfortunately) creates a bit of a special case here for this error
condition as we no longer have an "error message" to display to the user.  So,
instead of treating it as an "error condition", just redirect the User to the
login form (as presumably seeing the login form should indicate to the User
that they need to login, and any note saying "You need to login" is just
pointless).