Fix {bz: 3761}; when we guess the URL to redirect you to on login, some URLs just shouldn't be redirected to.
When the Challenger gets re-triggered because of a login failure (e.g. you
provided bad creds from a login page), _and_ we weren't provided any URL to
redirect you to, it uses "the current URL" as a guess for where we think you
wanted to end up.

However... when we weren't provided with a redirect URL in the first place, we
end up guessing "/nlw/submit/login", which just creates an infinite loop; once
you've logged in, you're redirected _back_ to the login page again and again.

So.... watch out for a small handful of known bad/icky URLs for "the Challenge
system and our STLogin form" and if we ever end up guessing (or are handed)
those then un-guess that and do something far more sane; leaving '$redirect'
as undef will cause us to eventually redirect to "/" (which is the sanest
thing we could do here).