{bz: 3809} Make plugin_enabled_for_user_sets do what it's name suggests.
It would only work if both users were user_ids or if the 'user_set_id'
parameter was a group.  When passing in an account (which isn't contained in
anything except for AUWs) the query would fail.  The SQL now handles this
case.  Indirect coverage is handled by the signal-reply wiki tests.