Fix "challenge" redirects, so we don't break WS's starting with the word "challenge".  Fixes Bug {bz: 3911}
Tighten up the regexes so that we don't catch WS's that start with the word
"challenge".  We're still unable to have a WS with the _name_ "challenge", but
its now ok to have one that starts with that word (e.g.
"challenge-everything").

(cherry picked from commit 2db557a382481160ca5385b4e7dfe276385258fa)