New wikiD tests for Bug {bz: 3911}.
Test to ensure that its ok to have a WS that begins with the word "challenge"
(and that the fix for that doesn't accidentally break triggering the
challenger when we really want to).

Two separate wikiD tests; one for "regular" browsers, the other for "mobile" ones.

(cherry picked from commit a4002920d514a27199f6203351bb0766ef0e5b51)