Fix {bz: 1260} by utf8-encoding addresses.
By not encoding the to: address, a unicode-containing username was causing the
entire header string-scalar to be flagged as unicode (SVf_UTF8).  When the
header was then concatenated to the body - which had already been encoded to
bytes from unicode - the resultant string was infected with the unicode flag.
This resulted in the correct encoding of the header portion, but the
double-encoding of the body portion.