Challenger which gets User info from Client-Side SSL certificate.
Although the "SSL Cert Creds Extractor" is nice, it _doesn't_ set an NLW-user
cookie, which is required for the JSON Proxy to work properly.

So... use a custom Challenger instead, which grabs the creds that were passed
in from the front-end about the SSL Certificate, and then use that to set an
NLW-user cookie and redirect the User back to where they wanted to be in the
first place.