WS Admins should get shown instructions on how to create a new WS.
Previously, WS Admins could create their own WS from the WS Settings page.
We've now changed this so that they have to create a Group first and then
create the WS for the Group.

Needless to say, some Users got confused at the change in behaviour.  This
commit shows a page of instructions to the WS Admin telling them how to go
about doing this.