Compensate for a bug in Socialtext::WikiText WRT a-phrases.
The "array of regexes" grammar rules don't quite work how we want in the base
class; we don't want "whichever regex matches first in the array" but
"whichever regex in the array matches closest to the front of the input
string".  Fortunately, we can provide a `match_a` method or single regex that
would do this.  Since luke and I couldn't figure out a single regex that
*also* retained the $1 and $2 captures, we chose to use the method instead.