Fix handling of htmldoc error conditions.
htmldoc doesn't return 512 or 1024 when there's an error... the return code is
based on the number of errors that were detected during the generation of the
PDF (shifted to the right by 8bits as the RC) (e.g. 2 errs == 512)

htmldoc also has a habit of trying to do HTTP lookups even when you give it
file paths, each one generating an error if it can't do the lookup.  Thus, for
a page with one image attached to it we *expect* to get 2 errors; one for
failed HTTP lookup for the page, one for the failed HTTP lookup for the image.

Rework the error handling so that we _ignore_ these error messages but treat
*any* other error message as indication of an error.

Pleasantly enough, for pages that contained "one" or "three" images, this now
means that we _aren't_ retrying five times but only generate the PDF once.
For PDFs matching this criteria, this cuts 80% of the overhead in generating
the PDF.  Unexpected perf bonus! :)