Factor out the wiki->html conversion for the page.
Its looking like part of the problem in Bug {bz: 3915} is a periodic failure
to convert the wikitext to HTML (which then manifests itself as an htmldoc
error as it doesn't have any HTML to convert to PDF).

Factoring this out into a var will make it a bit easier to poke at, and also
limits the scope of the "no warnings" statement here.