Output a better error message when we fail to export to PDF.
When htmldoc fails to convert a page to PDF, we don't get any results back
from it, so why were we thinking we can bubble that out to the end-user?

Errors are now actually logged, so output a more reasonable error message to
the User to let them know that there was a problem.