Do a better job of capturing/logging output from htmldoc.
htmldoc generates output that really wants to go to a TTY, so its got "\r"s in
it. :|

So... when we grab all of the error output, split it up properly so that we
get to see ALL of it in the logs.

And... make sure that we log the *full* error output when something goes wrong
(not just the error messages themselves); if we really have a problem we're
going to find the full context here useful.

As for dev-envs, *always* output the htmldoc output.  Even when we don't have
errors, it'll be worth logging so we can see what's going on.