Only grab Group Ids.  Allows for total purge of Groups, regardless of whether they can be instantiated or not.
This is related to Bug {bz: 3720}; if you've got LDAP Groups in the system and
the LDAP store goes away, those Groups can't currently be instantiated.  If
you can't instantiate the Group, you can't delete it.

Here, though, we don't care if LDAP is up/down/whatever... we just want to get
rid of *ALL* of the Groups.