Upgrade job to index Office 2007 Page Attachments.
Hunts down all of the Office 2007 attachments that exist in a WS and creates a
pile of AttachmentIndex jobs to go index each of those in the background.

Goal is to end up the Upgrade job quickly (so we don't block multi-version
upgrades for days/weeks at a time).  The individual attachment index jobs are
given _lower_ priority than normal, so that they don't clog the queue.
*Eventually* they'll all get done/re-indexed, but that doesn't have to be
'right now'.