Make it obvious when we restrict weblog display limit.  For {bz: 4120}
Although we were properly restricting the weblog display to the max limit, it
wasn't obvious that this was taking place.  Add some logging so that we leave
a breadcrumb trail for ourselves (so next time we think we're serving up
grillions of weblog entries at once we'll have something in the logs to remind
us that we have in fact limited the results).