Force auto creation of People Profile Fields.
Now that we've got some extra fields coming from LDAP to go into a People
Profile, we need to make sure that we've got a full set of People Fields ready
and available to load the values into.  We can either do it the long way by
creating an Account, enabling the People plugin, and then creating the fields
in the Account, *or* you can just turn on this flag. :)