Fix {bz: 4195} (1 of 2) fix refresh algorithm.
Using DN to determine the membership set of a flattened group is bad:
1) DN isn't necessarily case-sensitive
2) a DN can be a referral to another object that *is* in the group.

Using the user_id number of each user, we're able to track the exact set and
avoid not only the add_user() error, but other related errors as well.

During testing it was also discovered that we don't handle escaped characters
in groups very well, so this patch also includes that fix.