When configured "ssl_only", *require* that an SSL cert be available.
If the appliance is configured to run in "ssl_only" mode, we *NEED* to have an
SSL certificate available for use (so fail/abort if its not there).

Normally, presence of an SSL certificate is optional, but if we know that
we've been explicitly configured to need/require one, then it sure better be
there.