Mark LDAP User as missing/found based on whether our lookup was successful.
If we _used to_ have a record for the User but can't find them in LDAP any
more, mark them as "missing".  That way we know that we _tried_ to do a lookup
but just weren't able to find them.

If the User comes back again, mark them as "found" (not missing).
(cherry picked from commit 31580132b86cdf6b8fc4544af9158cfb9e6e7828)