Bump "cached_at" when User continues to be found as "missing".
If we don't bump this ahead, we'll start to query the User on every lookup
again once we reach the "not_found_ttl".  By bumping it ahead, we ensure that
we're only doing one lookup per TTL period.

Tests updated to do multiple lookups so we can verify that we only went out to
LDAP the minimal amount of times.