If the User is "missing", create an actual ST:U:Deleted homunculus.
When we find out that a User is missing from LDAP the very first time, we'd
create a ST:U:Deleted homunculus to reflect that.

When pulling up _cached_ data from the DB, we need to ensure that we're *also*
creating the same type of Homunculus object; ST:U:Deleted has over-rides to
several methods and attributes and we need to ensure that this behaviour is
preserved even when using cached data.
(cherry picked from commit 68e37f7c8bb18c5866c5a118f2fabbf247e5dfcc)