Ensure that LDAP User data is properly UTF8 decoded.
Fixes "deep recursion" errors when trying to refresh LDAP Users that have UTF8
names.

Although it appears that we're properly marshalling things to/from the DB,
when a DB value is compared to what we got back from LDAP I'm seeing that one
is properly utf8 decoded while the other is not.