Fix {bz: 4344}.  Don't skip adding LDAP Users if they were previously known.
If a User was added to an LDAP Group that was _already_ known to the
Socialtext system, we were accidentally skipping over them when updating the
Group membership list during refresh.

Users that we'd never seen before were properly added, but Users that we
already knew about were skipped over by accident.