Fix {bz: 4373} - Don't auto-update LDAP Users, even if "auto-provisioning" is enabled.
Even if auto-provisioning is enabled, *do not* attempt to update an LDAP
sourced User; those Users auto-update automatically by virtue of being LDAP
sourced.

We do, however, need to continue to be able to auto-provision _other_ Users,
though, as we may be operating in a mixed-mode environment (some LDAP Users,
some Default ones).