Add a new Account User role.
The Account User is a user that is 1) Authenticated, 2) shares an account with
the current Workspace, and 3) is _not_ already a Member or an Admin.

This role is a "virtual" role, like Guest and Authenticated User; the role is
not explicitly assigned to a user by an admin. Rather, a series of
computations takes place to determine whether or not a user is an Account User
for a given workspace.

This role lets us assign permissions to user's that share an account with the
workspace, but do not otherwise have an assigned role.