Cache cookie parsing/decryption, for perf.
No point decrypting the same cookie again and again and again; if we've done
it once and got back a token just cache it as we'll get the same results next
time too.

Checks against the _contents_ of the cookie aren't cacheable, but the actual
decryption of the cookie is.