Resolve race-condition in WorkQueue.pm.
The implementation is changed from a linked-list to using Coro::Channel
directly. This gets rid of the race and should make this module a bit more
reliable. The coro working the jobs still cede()s, which is an important
feature of this module (but hard to cover in unit tests).

Coro::Channel wasn't used because I thought (and heard from some guy at YAPC)
that using an array like this leaked memory.  In experiments, it doesn't seem
that the array Coro::Channel is using leaks memory after all.  The mem *does*
remain allocated for an empty queue, however, so just hack that case to reset
the array.  A similar approach is used in the Socialtext::PerfAgent code.