Helper method to get the full list of _all_ the headers needed for Creds Extraction.
We'd like to optimize what we send from the CE Client to the st-userd
back-end, so we're not sending along the full set of inbound HTTP headers.
So... ask the CEs which headers they actually need, so we can optimize what we
send across the wire to the smallest possible set of headers.