New PersistentHash cache back-end.  For times when you *don't* ever want the cache to be cleared.
Over the years we've adopted a pattern of expecting that the in-memory caches
are cleared between HTTP requests.  There are times, though, when we have a
cache that we *know* we want to have persist across requests.

In those cases, this back-end ignores requests to clear the cache and thus
persists across requests.