Fix {bz: 4509} by using a correct table-lock mode.
The transitive-closure algorithm used for the UserSet tables requires that
there are no concurrent updaters.

Locking a table in just SHARE mode doesn't prevent concurrent updaters.  This
was my mistake, originally.  Fortunately, this sort of lock (the fact that the
code is locking two tables at once) was always resulting in a deadlock,
preventing the concurrent update (and thus data corruption).

Under the SHARE ROW EXCLUSIVE mode, concurrent processes executing the LOCK
statement will block until that lock is obtained.  Thus, the updates are
serialized and don't spew deadlock errors.  Readers should still be able to
access the table while it's updating.