Force a minimum 24h limit for auth tokens, *IF* they try to set to <=0
We've had at least one customer try to set this to <=0 forcably in the config
file, so try to prevent them from doing that.

We still want to allow for (potentially insanely) low values for testing, but
don't let them set it to a point of total and utter breakage.