Fix {bz: 4542}; set WS perms _before_ firing import hook.
If we fail to set the permission-set for the WS before firing off the
"nlw.import_workspace" hook, any plugins that get invoked will be working
against a WS that isn't really set up yet.

The "Groups" plugins specifically has problems with this, as it could
incorrectly detect a mismatch for the WS/Group permission sets (as the WS
would still be marked as being "member-only", and that conflicts with a
"self-join" Group even if the WS is eventually going to be "self-join").