Build SQL for UserMetadata creation on-the-fly, don't hardcode the SQL statement.
Allows for *all* of the fields in the UserMetadata table to be specified,
including "last_login_datetime" and "creation_datetime".

Without this, we lose that information on across User export/import; even
though the data is present in the export it wasn't being properly imported
back in on the other side.