Optionally refetch Groups from the DB when they're created.
Some tests require that we have the actual/exact record from the DB, as
opposed to the initial in-memory copy that we created.

Most notably, tests needing to compare the "creation_datetime" of the Group
make use of this, as DateTime->now(hires=>1) has a different resolution than
when it gets pulled up out of Pg; if you try to compare one to the other
you'll be off.