Add a means to test large account exports.
This will become a means for adding/maintaining account exports with large
amounts of data. The goal here is to purposely make things as messy as
possible so we can sniff out corner case failures with account export/import.

Obviously, there's still fair amount of work that needs to be done in order to
build up this framework.

Include a script that will reproduce the messiness of the account in a
dev-env. If it's good enough to detect account import/export corner cases,
it'll probably be useful for finding corner cases elswhere in our code.