First pass at a creds extractor for client-side ssl certificates.
Nginx is configured to pass through the Client Subject DN via an HTTP header,
which we then pick apart to get the Username out of.

First pass ... only grabs it from "CN" and doesn't accommodate embedded "/"s
in the DN when we convert "/" delimited DNs to "," delimited ones (so we can
leverage some Net::LDAP code to extract the fields for us).