Put the private fields in our "minimal_interface", but grep them out.
If they're not in the minimal interface, we're not going to be able pull them
back in on User import easily.

Fortunately, the *only* place that `@minimal_interface` is used is in the
"to_hash()" and "Create_user_from_hash()" methods, so we're not risking having
these fields exposed out any further than just here.