Bump SiteMinder log msgs from "debug" to "info".
While working through SiteMinder integration issues recently with one of our
customers, we found it incredibly helpful to have the additional logging.

For the common case of "Users having proper access", **NO** additional logging
is performed.

For the case of Users having trouble accessing the system, though, additional
information now appears in the system logs to assist in debugging the problem
that the User(s) are having.