Mark LDAP Users as having been refreshed *before* validating the data we pulled from LDAP.
If validating the data or saving the data causes a hook/trigger/plugin to get
fired that needs access to the User object, we want to prevent endless looping
while trying to re-instantiate the User record again and again.

Mark it as being cached, and the next guy through will think that what we've
got in the DB is "good enough" for now.  At the end of *our* refresh here,
when *we* save the User record, those hooks/plugins should get triggered
again, causing the *updated* data to get used.