Use the "guess_real_name" as the "display_name" for Users.
In order to ensure that we properly sort and display Users by their Preferred
Name, the PN needs to be stored *in* the User record.

The "display_name" field in the DB exists already for this purpose (having a
place in the DB where we store the name for sorting/display purposes), so just
change its calculation so that it takes into consideration a User's "Preferred
Name" (if they have one).