Its ok for this to be an actual Homunculus now.
Now that we're updating the "cached_at" properly to prevent infinite recursion
while trying to update LDAP Users, we can actually instantiate a full
Homunculus here.

That's good, because we need a *real* Homunculus object to validate the LDAP
data against (in case we need to pull the Preferred Name for the User).