There is *no* factory that you can use to create a Deleted User object.
Deleted Users have no factory to create them.  Subsequently, there isn't a way
to go back and update them cleanly.

"Deleted Users" are also the exception to the rule architecturally to how our
User subsystem is built, so we'll keep this exception where it belongs, *in*
the module with the other parts of the exception (rather than pushing it
outwards and polluting ST:User:Base).