Method to update the "display_name()" for a User record, *directly* in the DB.
LDAP Users aren't considered to have an updateable store, so we can't use
"$user->update_store()" to change the "display name" for those Users when any
of the other information on the User changes.

Instead, we go direct to the Factory and ask it to update the underlying DB
record to do the work.