*Always* recreate the display name when validating User data.  Always.
Even if we don't have a first/last/email, there are other things beyond the
scope of the User's data that could affect the display name (e.g. People
Profile fields).  If those values change out there, we need to ensure that the
display name gets updated properly for the User.