Re-fix {bz: 4857} (part 2 of 2); enable general User cache so Users only get looked up once.
Although we want the initial lookup to go out and refresh from LDAP, we only
want to look that User up once.  Enable the general User cache to ensure that
if we end up chasing relationships that we're not recursing.

Cache is cleared between User lookups to ensure that each User being refreshed
does in fact get out to LDAP for refresh.