Patch run_in_txn to work with nested transactions.
In order for savepoints to work, we need to wrap the inner TS:M code in a
sql_txn block.  If we don't do this, and a job-insert fails and causes the
transaction to cancel, the outer savepoint/transaction is effectively
cancelled.  TS:M doesn't attempt to make a nested transaction (but simply
checks to see if one is active already), which is what caused this undesirable
behaviour.
(cherry picked from commit a618d21f94c74a056df2fca605f6ef859de6a1b6)