Fixes {bz: 4937} - st-daemon-monitor signals more widely
we believe supervise was dying, and then losing track of the PID of the
st-userd that was actually bound to the port, and thus
st-daemon-monitor was not sending a signal to that st-userd process that was
listening on the port (8084). st-daemon-monitor now *also* does a killall
-$signal $procname to ensure that all processes (whether or supervise has the
right $PID) get the signal.
(cherry picked from commit a1e6d4ca3244448cc7da6aec692bed6f92e5f9ff)