Moar Indexes!
The workspace name one is speculative; seems like it could be used for
workspace name lookahead.

The page_revision ones are all required for efficient lookup.  Add duplicates
with `NOT deleted` predicates since most queries will add that (optimize).

The page ones add a last_edit one for parity with create, and predicated ones
for non-deleted pages (optimize).

The page_tag ones are optimizations for lookahead and workspace-listing.  The
(tag) and (workspace_id) ones weren't used as far as i could tell.