Unwind some of the helper methods used to work with e-mail confirmations.
Reduces the number of function calls we have to make, eliminating the need for
some of the caching that was put in place around the Email Confirmations.

No point dipping into the User object for all of this stuff time and time
again, when we could just grab the Confirmation object and then ask it all the
questions that we have; going through the User object is just a whole level of
indirection that's just overhead.