Make this method return the Restriction, not the User.
No point using the hash/token to get the Restriction and return the User, as
we're then using the User to go get the Restriction again.  Already had the
Restriction in our hand, just return it to the caller.  If they want the User,
they can ask the Restriction for it.