Users with *any* outstanding restrictions cannot be logged in.
More than just having an outstanding "email confirmation", we now want to make
sure that the User cannot be logged in if they have _any_ outstanding
restrictions (which would be either an "email confirmation" or a "password
change").

These restrictions used to be merged together into a single row in the DB that
did double-duty, but they've since been split out into separate records so
they can be handled individually.