Cheater method to update the External Id for a User record.
Have to cheat a bit here as it'd be possible to want to update this for an
LDAP User record (and LDAP Users are treated as non-modifyable).

Yes, this would be better done as a refactoring to the Homunculus and Factory
classes so that they allowed for internal fields to be updated on LDAP User
records, but that's beyond the scope of what we're doing here right now.