Fix up DB/object mapping, when creating a new Homunculus object.
This one keeps coming back and biting us again and again, as its named
different in the DB than it is in the object (and we're _regularly_ converting
it from one to the other).

Push this fix-up in down here, which should let us not have to think about it
at a higher level.