Fix {bz: 5019}, by removing a DB lookup for *every* User object.
User objects were being looked up twice in the DB, when looked up by
"user_id"; one lookup was done to figure out which User Factory was
responsible for the User, and then the Factory itself would do a second lookup
to get the User data from the DB.

To avoid the second lookup, pass the results from the first one through as a
"preload => $row", so that the Factory knows that we've already done a lookup
and have the data we need here.

Default User Factory then just instantiates directly from that (as that's all
it was doing before-hand).

LDAP User Factory gets to bypass its cache lookup, and uses this preloaded
copy of the data as that lookup.