New "sslversion" LDAP config option, to *force* a specific SSL version.
If you need to *force* a specific SSL version to be used when connecting to an
LDAPS server, this option allows you to do so.

By default, Net::LDAP attempts to connect using "tlsv1", but if your LDAPS
server doesn't support that then you may need to force a specific version.