Use a common Base DN to hold the root user/pass in, allowing for multiple DBs, which we need for LDAP Referral testing.
OpenLDAP-2.4.x no longer supports issuing of referrals for a Base DN that it
has a listed database for.  In order to get referrals working properly again,
we need a main database entry to hold the root user/pass in which is
*separate* from whatever database we're using to hold our actual LDAP data in
for testing.

To support this further, a "nodb => 1" option can be passed in, to *disable*
the configuration of a database for the LDAP test data.  You'll be left with
the common Base DN with the root user/pass in it, but that's it.