TablePopulator now accepts a 'skip_user_create' param.
If that's passed, we will _not_ attempt to create a new user when we come
across a page edit/attachment that contains an unknown user. Instead, we will
assign that edit/attachment to the SystemUser, keeping a log so we can go back
and clean things up later if we'd like.

This should never be needed in the normal course of operations, but it'll come
in useful for doing imports concurrently, such as when we use the big
migration tool.