Fix {bz: 5117}; allow for a Private Group to be created with the same name as an LDAP Group.
Key here is that the Group's "unique-ness" isn't on "name", but is on
"name+account+creator".  It *is* possible to have multiple "Golfing Buddy"
Groups in the system, provided that they've got different creators.  Its
also possible to have a private/built-in Group with the same name as an LDAP
Group; don't restrict against it.