Make the "reports" and "console" tests run again.  (but see caveats below)
These tests (unfortunately) require that they be run in their original trees,
and *cannot* be symlinked into the main test tree in their current state; the
tests make *extensive* use of mocks, and expect to be run from outside of the
main source tree.

This patch gets the tests running, **BUT** we don't get any of the state
information stored in ".prove", as the state for these tests is in the
".prove" files in those other out of tree repos.