Don't merge test results into .prove
Problems stems from our then ending up with tests from other repos being
bundled in.  Although this would be nice as it'd bring that info over from the
other repos, the problem is that when we then re-run failing tests to confirm
that they're not spurious failures, that we try to run -those- tests again and
we end up erasing the TAP output from the initial run.