Fix {bz: 5321}. This represents a major change in how LDAP works.
Previously, the DB and the LDAP user store(s) operated with no knowledge
whatsoever of one another. This worked fine when the different stores were in
sync. When one began to diverge from the other, all hell would break loose.

A major contributing factor here is that the users.driver_username field in
the DB was not considered unique. While the field is still not unique (look
for that in a further commit), this code does its best to update user records
rather than create new records.