Early out if the user is deleted.
If we find a deleted user in the DB just return instead of
trying to reconcile in LDAP. This is change in functionality as before
we would search all LDAP stores for a potential match. Since the username
should be unique for the user, there is no need to do that.

If the user really should be found, toggling the is_deleted flag on the user
record will result in the right thing happening.