Fix an attachment "cleanup" bug.
Occasionally, there is a job that runs which deletes attachments that are not
referenced (read foreign key) anywhere else in the DB. Because account themes
are stored as blobs in the DB, we can't possibly set up an FK constraint
there. Thus we need to create a link table, account_theme_attachments.